Refactor login functionality and enhance UI components
- Introduced a new LoginPage component with separate sections for email and QR code login.
- Added styled components for better UI structure and responsiveness.
- Implemented error handling and loading states for login processes.
- Updated logo positioning logic based on window size.
- Refactored existing login logic into reusable components for improved maintainability.
- Removed deprecated LoginPage.tsx file to streamline the codebase.
