+++ frontend/src/app/hooks/useWindowSize.ts
@@ -1,11 +1,18 @@
 import { useLayoutEffect, useState } from 'react';
 
 export const useWindowSize = () => {
+  const isServerSide = typeof window === 'undefined';
+
   const [size, setSize] = useState({
-    width: window.innerWidth,
-    height: window.innerHeight,
+    width: isServerSide ? 1920 : window.innerWidth, // Default to common desktop size on server
+    height: isServerSide ? 1080 : window.innerHeight,
   });
+
   useLayoutEffect(() => {
+    if (isServerSide) {
+      return;
+    }
+
     function updateSize() {
       setSize({
         width: window.innerWidth,
@@ -16,6 +23,7 @@ export const useWindowSize = () => {
     window.addEventListener('resize', updateSize);
     updateSize();
     return () => window.removeEventListener('resize', updateSize);
-  }, []);
+  }, [isServerSide]);
+
   return size;
 };
