+++ frontend/src/app/pages/login/components/ErrorMessage.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const ErrorMessage = styled.div`
+  color: #ff6b6b;
+  font-size: 14px;
+  margin-top: 10px;
+  text-align: center;
+`;
