+++ frontend/src/app/pages/login/utils/formatCode.tsx
@@ -1,0 +1,3 @@
+export const formatCode = (code: string): string => {
+  return code.slice(0, 4) + ' ' + code.slice(4);
+};
