+++ frontend/src/app/pages/login/utils/formatTimeRemaining.tsx
@@ -1,0 +1,6 @@
+// Format time remaining as MM:SS
+export const formatTimeRemaining = (seconds: number): string => {
+  const minutes = Math.floor(seconds / 60);
+  const remainingSeconds = seconds % 60;
+  return `${minutes}:${remainingSeconds.toString().padStart(2, '0')}`;
+};
